// apps/mobile/src/api/generated/models/reasoningSelectionDto.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * 办伴 Banban API
 * 办伴 Banban 后端接口文档
 * OpenAPI spec version: 1.0
 */
import type { ReasoningSelectionDtoActivation } from './reasoningSelectionDtoActivation';
import type { ReasoningSelectionDtoEffort } from './reasoningSelectionDtoEffort';

export interface ReasoningSelectionDto {
  activation?: ReasoningSelectionDtoActivation;
  effort?: ReasoningSelectionDtoEffort;
  budgetTokens?: number | 'auto';
}